+++ src/struphy/models/incompressible_navier_stokes_sph.py
@@ -17,7 +17,7 @@ from struphy.propagators.base import Propagator
 from struphy.propagators.poisson_solve import PoissonSolve
 from struphy.propagators.push_eta import PushEta
 from struphy.propagators.push_v_in_force_field import PushVinForceField
-from struphy.propagators.push_vin_viscous_potential import PushVinViscousPotential
+from struphy.propagators.push_v_in_viscous_potential import PushVinViscousPotential
 from struphy.propagators.push_vxb import PushVxB
 
 
@@ -217,7 +217,7 @@ class IncompressibleNavierStokesSPH(StruphyModel):
 
         1. :class:`~struphy.propagators.push_eta.PushEta`
         2. :class:`~struphy.propagators.push_vxb.PushVxB` (if :attr:`with_B0` is True)
-        3. :class:`~struphy.propagators.push_vin_viscous_potential.PushVinViscousPotential` (if :attr:`with_viscosity` is True)
+        3. :class:`~struphy.propagators.push_v_in_viscous_potential.PushVinViscousPotential` (if :attr:`with_viscosity` is True)
         4. :class:`~struphy.propagators.poisson_solve.PoissonSolve`
         5. :class:`~struphy.propagators.push_v_in_force_field.PushVinForceField`
         """
