+++ src/struphy/propagators/push_v_in_viscous_potential.py
